refactor: decouple main.py into service and endpoints (fixes #227)
